functions/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
google-cloud-pubsub
Wand
django-environ
requests
openpyxl
google-api-python-client
google-auth
